Release/0.9.0 (#86)
* Sync with platform api 2025 09 22 (#75)

* OpenAPI Changes

* Generated provider and docs

* fix: tests

* OpenAPI changes to fix tests

* Generated provider

* OpenAPI changes - after ignoring some mesh changes

* Generated provider and docs

* tests: fix tests

* Update CHANGELOG.md

* Feat/event gateway without policies (#85)

* OpenAPI changes

* Generated provider and docs

* tests: add and skip tests

* chore: prepare for v0.9.0 release

* Update changelog with mesh features

* tests: skip failing mesh policy test

+++ internal/provider/types/endpoints.go
@@ -8,5 +8,5 @@ import (
 
 type Endpoints struct {
 	Address types.String `tfsdk:"address"`
-	Port    types.Int64  `tfsdk:"port"`
+	Port    types.Int32  `tfsdk:"port"`
 }
